DataStructuresAndAlgorithms/src/main/java/me/zhang/dataStructure/LinkedStack.java
package me.zhang.dataStructure;

/**
 * Created by zhangxiangdong on 2017/6/2.
 */
public class LinkedStack<E> implements Stack<E> {

    private SinglyLinkedList<E> list = new SinglyLinkedList<>();

    @Override
    public void push(E e) {
        list.addFirst(e);
    }

    @Override
    public E pop() {
        return list.removeFirst();
    }

    @Override
    public boolean isEmpty() {
        return list.isEmpty();
    }

    @Override
    public E top() {
        return list.first();
    }

    @Override
    public int size() {
        return list.size();
    }

}
